chore: idea for a refactor/extraction
