+++ packages/jhehemann/skills/kelly_strategy/README.md
@@ -1,0 +1,5 @@
+# Kelly Strategy
+
+## Description
+
+This skill implements a number of helper methods that implement the kelly strategy. They serve as plugins for the decision_maker_abci of the trader agent.
