+++ packages/jhehemann/skills/kelly_strategy/skill.yaml
@@ -1,0 +1,20 @@
+name: kelly_strategy
+author: jhehemann
+version: 0.1.0
+type: skill
+description: A degenerate skill providing helper functions for the kelly strategy to the decision_maker_abci of the trader agent.
+license: Apache-2.0
+aea_version: '>=1.0.0, <2.0.0'
+fingerprint:
+  README.md: bafybeia367zzdwndvlhw27rvnwodytjo3ms7gbc3q7mhrrjqjgfasnk47i
+  __init__.py: bafybeih563ujnigeci2ldzh7hakbau6a222vsed7leg3b7lq32vcn3nm4a
+fingerprint_ignore_patterns: []
+connections: []
+contracts: []
+protocols: []
+skills: []
+behaviours: {}
+handlers: {}
+models: {}
+dependencies: {}
+is_abstract: true
