+++ tox.ini
@@ -64,7 +64,7 @@ setenv =
     PYTHONPATH={env:PWD:%CD%}
     PACKAGES_PATHS = packages/valory
     SKILLS_PATHS = {env:PACKAGES_PATHS}/skills
-    SERVICE_SPECIFIC_PACKAGES = {env:SKILLS_PATHS}/staking_abci {env:SKILLS_PATHS}/market_manager_abci {env:SKILLS_PATHS}/decision_maker_abci {env:SKILLS_PATHS}/trader_abci {env:SKILLS_PATHS}/tx_settlement_multiplexer_abci
+    SERVICE_SPECIFIC_PACKAGES = {env:SKILLS_PATHS}/staking_abci {env:SKILLS_PATHS}/market_manager_abci {env:SKILLS_PATHS}/decision_maker_abci {env:SKILLS_PATHS}/trader_abci {env:SKILLS_PATHS}/tx_settlement_multiplexer_abci packages/jhehemann/skills/kelly_strategy
 
 [testenv:bandit]
 skipsdist = True
